refactor: move chain-storage-paths to internal
